translator-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai>=1.0.0
httpx
